# openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ExecuteRequest": {
        "properties": {
          "variables": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Variables",
            "type": "object"
          }
        },
        "title": "ExecuteRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedResponse": {
        "description": "Standard envelope for list endpoints.",
        "properties": {
          "items": {
            "items": {},
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items"
        ],
        "title": "PaginatedResponse",
        "type": "object"
      },
...
      "RequestCreate": {
        "properties": {
          "authConfig": {
            "additionalProperties": true,
            "title": "Authconfig",
            "type": "object"
          },
          "authType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authtype"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "bodyType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodytype"
          },
          "collectionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collectionid"
          },
          "headers": {
            "items": {
              "$ref": "#/components/schemas/ParamDto"
            },
            "title": "Headers",
            "type": "array"
          },
          "method": {
            "default": "GET",
            "title": "Method",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "queryParams": {
            "items": {
              "$ref": "#/components/schemas/ParamDto"
            },
            "title": "Queryparams",
            "type": "array"
          },
          "url": {
            "default": "",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "RequestCreate",
        "type": "object"
      },
      "RequestResponse": {
        "properties": {
          "authConfig": {
            "additionalProperties": true,
            "title": "Authconfig",
            "type": "object"
          },
          "authType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authtype"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "bodyType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodytype"
          },
          "collectionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collectionid"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "headers": {
            "items": {
              "$ref": "#/components/schemas/ParamDto"
            },
            "title": "Headers",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "projectId": {
            "title": "Projectid",
            "type": "string"
          },
          "queryParams": {
            "items": {
              "$ref": "#/components/schemas/ParamDto"
            },
            "title": "Queryparams",
            "type": "array"
          },
          "revision": {
            "default": 1,
            "title": "Revision",
            "type": "integer"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "method",
          "url"
        ],
        "title": "RequestResponse",
        "type": "object"
      },
      "RequestUpdate": {
        "properties": {
          "authConfig": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authconfig"
          },
          "authType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authtype"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "bodyType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodytype"
          },
          "collectionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collectionid"
          },
          "headers": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ParamDto"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headers"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "queryParams": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ParamDto"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queryparams"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "title": "RequestUpdate",
        "type": "object"
      },
      "RunResultResponse": {
        "properties": {
          "body": {
            "default": "",
            "title": "Body",
            "type": "string"
          },
          "durationMs": {
            "default": 0,
            "title": "Durationms",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "statusCode": {
            "title": "Statuscode",
            "type": "integer"
          }
        },
        "required": [
          "statusCode"
        ],
        "title": "RunResultResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first API workflow studio",
    "title": "SCLPLAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/collections/{collection_id}/requests": {
      "get": {
        "description": "List all requests in a collection.",
        "operationId": "list_collection_requests_api_v1_projects__project_id__collections__collection_id__requests_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "title": "Collection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Collection Requests",
        "tags": [
          "collections"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/requests": {
      "get": {
        "description": "List requests, optionally filtered by collection.",
        "operationId": "list_requests_api_v1_projects__project_id__requests_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "collection_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Collection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Requests",
        "tags": [
          "requests"
        ]
      },
      "post": {
        "description": "Create a new request.",
        "operationId": "create_request_api_v1_projects__project_id__requests_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Request",
        "tags": [
          "requests"
        ]
      }
    },
    "/api/v1/projects/{project_id}/requests/{request_id}": {
      "delete": {
        "description": "Delete a request.",
        "operationId": "delete_request_api_v1_projects__project_id__requests__request_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Request",
        "tags": [
          "requests"
        ]
      },
      "get": {
        "description": "Get a single request.",
        "operationId": "get_request_api_v1_projects__project_id__requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Request",
        "tags": [
          "requests"
        ]
      },
      "patch": {
        "description": "Update a request. Supports revision-based conflict detection.",
        "operationId": "update_request_api_v1_projects__project_id__requests__request_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Request",
        "tags": [
          "requests"
        ]
      }
    },
    "/api/v1/projects/{project_id}/requests/{request_id}/execute": {
      "post": {
        "description": "Execute an HTTP request and return the result.\n\nAlso saves the result to history. Loads the active environment's\nvariables and merges them with explicitly-passed variables.",
        "operationId": "execute_request_api_v1_projects__project_id__requests__request_id__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ExecuteRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Request",
        "tags": [
          "requests"
        ]
      }
    },
    "/api/v1/projects/{project_id}/requests/{request_id}/history": {
      "get": {
        "description": "List execution history for a specific request.",
        "operationId": "list_request_history_api_v1_projects__project_id__requests__request_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Request History",
        "tags": [
          "requests"
        ]
      }
    },
    "/api/v1/projects/{project_id}/requests/{request_id}/move": {
      "post": {
        "description": "Move a request to a different collection (or unassign).",
        "operationId": "move_request_api_v1_projects__project_id__requests__request_id__move_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_collection_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Collection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Move Request",
        "tags": [
          "requests"
        ]
      }
    },
...
  }
}